requirements.txt
numpy==1.24.2
moviepy==1.0.3
Pillow==9.5.0
